Update seeds
* Calling `dev:prime[EMAIL]` will create a user for email and make it an
  admin
* This task clears out DB so there are no problems with uniqueness
* Creates all records necessary to preview emails

https://trello.com/c/voBTTwVx
