+++ .env.example
@@ -4,6 +4,7 @@
 #
 MYUSA_KEY: consumer_public_key
 MYUSA_SECRET: consumer_secret_key
+DEFAULT_URL_HOST: your_local_server_address
 
 # optional
 #
@@ -11,7 +12,6 @@ MYUSA_SECRET: consumer_secret_key
 # BULLET_ENABLED=true
 # BUDGET_REPORT_RECIPIENT
 # DATABASE_URL
-# DEFAULT_URL_HOST
 # DEFAULT_URL_PORT
 # DEFAULT_URL_SCHEME
 # DISABLE_CLIENT_SLUGS=foo,bar,goop
