+++ db/seeds.rb
@@ -1,9 +1,9 @@
 require_relative "chores/roles_conversion"
 
 # any commands in this file should be idempotent
-Role.find_or_create_by(name: 'observer')
-Role.find_or_create_by(name: 'client_admin')
-Role.find_or_create_by(name: 'admin')
+Role.find_or_create_by(name: "observer")
+Role.find_or_create_by(name: "client_admin")
+Role.find_or_create_by(name: "admin")
 
 # RolesConversion relies upon some ENV variables to set the initial User records.
 # it should be safe to run multiple times in the same environment, and to remove
