+++ lib/mail_previews/attachment_mailer_preview.rb
@@ -1,17 +1,11 @@
 class AttachmentMailerPreview < ActionMailer::Preview
+  include MailerPreviewHelpers
+
   def new_attachment_notification
     AttachmentMailer.new_attachment_notification(email, proposal, attachment)
   end
 
   private
-
-  def email
-    "recipient@example.com"
-  end
-
-  def proposal
-    Proposal.last
-  end
 
   def attachment
     Attachment.last
