+++ lib/mail_previews/cancelation_mailer_preview.rb
@@ -1,40 +1,30 @@
 class CancelationMailerPreview < ActionMailer::Preview
+  include MailerPreviewHelpers
+
   def cancelation_notification
     CancelationMailer.cancelation_notification(
-      recipient_email: email,
       canceler: user,
-      proposal: proposal,
-      reason: reason
+      proposal: canceled_proposal,
+      reason: reason,
+      recipient_email: email
     )
   end
 
   def cancelation_confirmation
     CancelationMailer.cancelation_confirmation(
-      proposal: proposal,
+      proposal: canceled_proposal,
       canceler: user,
       reason: reason
     )
   end
 
   def fiscal_cancelation_notification
-    CancelationMailer.fiscal_cancelation_notification(proposal)
+    CancelationMailer.fiscal_cancelation_notification(canceled_proposal)
   end
 
   private
 
-  def email
-    "test@example.com"
-  end
-
-  def user
-    User.last
-  end
-
   def reason
     "Example reason for canceling proposal"
-  end
-
-  def proposal
-    Proposal.last
   end
 end
