+++ lib/mail_previews/comment_mailer_preview.rb
@@ -1,15 +1,7 @@
 class CommentMailerPreview < ActionMailer::Preview
+  include MailerPreviewHelpers
+
   def comment_added_notification
     CommentMailer.comment_added_notification(comment, email)
-  end
-
-  private
-
-  def comment
-    Comment.last
-  end
-
-  def email
-    "recipient@example.com"
   end
 end
