+++ lib/mail_previews/step_mailer_preview.rb
@@ -1,4 +1,6 @@
 class StepMailerPreview < ActionMailer::Preview
+  include MailerPreviewHelpers
+
   def step_reply_received
     StepMailer.step_reply_received(completed_step)
   end
@@ -9,23 +11,5 @@ class StepMailerPreview < ActionMailer::Preview
 
   def proposal_notification
     StepMailer.proposal_notification(step)
-  end
-
-  private
-
-  def completed_step
-    Step.where(status: "completed").last
-  end
-
-  def step
-    Step.where(type: "Steps::Approval").last
-  end
-
-  def proposal
-    Proposal.last
-  end
-
-  def user
-    proposal.approvers.last
   end
 end
