+++ lib/mail_previews/welcome_mailer_preview.rb
@@ -1,11 +1,7 @@
 class WelcomeMailerPreview < ActionMailer::Preview
+  include MailerPreviewHelpers
+
   def welcome_notification
     WelcomeMailer.welcome_notification(user)
-  end
-
-  private
-
-  def user
-    User.for_email("test@example.com")
   end
 end
